adicionando rota update e delete

+++ src/database/migrations/20200724060338-createdados.js
@@ -19,7 +19,7 @@ module.exports = {
       allowNull: false,
     }, 
     participation: {
-      type: Sequelize.DOUBLE,
+      type: Sequelize.REAL,
       allowNull: false,
     }, 
     created_at:{
